+++ gctscript/modules/ta/indicators/mfi.go
@@ -7,17 +7,17 @@ import (
 	"strings"
 
 	objects "github.com/d5/tengo/v2"
-	"github.com/thrasher-corp/go-talib/indicators"
+	"github.com/thrasher-corp/gct-ta/pkg/indicators"
 	"github.com/thrasher-corp/gocryptotrader/gctscript/modules"
 	"github.com/thrasher-corp/gocryptotrader/gctscript/wrappers/validator"
 )
 
-// AtrModule range indicator commands
-var AtrModule = map[string]objects.Object{
-	"calculate": &objects.UserFunction{Name: "calculate", Value: atr},
+// MfiModule index indicator commands
+var MfiModule = map[string]objects.Object{
+	"calculate": &objects.UserFunction{Name: "calculate", Value: mfi},
 }
 
-func atr(args ...objects.Object) (objects.Object, error) {
+func mfi(args ...objects.Object) (objects.Object, error) {
 	if len(args) != 2 {
 		return nil, objects.ErrWrongNumArguments
 	}
@@ -32,9 +32,9 @@ func atr(args ...objects.Object) (objects.Object, error) {
 	if !valid {
 		return nil, fmt.Errorf(modules.ErrParameterConvertFailed, OHLCV)
 	}
-
 	ohlcvData := make([][]float64, 6)
 	var allErrors []string
+
 	for x := range ohlcvInputData {
 		t := ohlcvInputData[x].([]interface{})
 		value, err := toFloat64(t[2])
@@ -61,17 +61,15 @@ func atr(args ...objects.Object) (objects.Object, error) {
 		}
 		ohlcvData[5] = append(ohlcvData[5], value)
 	}
-
-	inTimePeriod, ok := objects.ToInt(args[1])
-	if !ok {
-		allErrors = append(allErrors, fmt.Sprintf(modules.ErrParameterConvertFailed, inTimePeriod))
-	}
-
 	if len(allErrors) > 0 {
 		return nil, errors.New(strings.Join(allErrors, ", "))
 	}
+	inTimePeriod, ok := objects.ToInt(args[1])
+	if !ok {
+		return nil, fmt.Errorf(modules.ErrParameterConvertFailed, inTimePeriod)
+	}
 
-	ret := indicators.Atr(ohlcvData[2], ohlcvData[3], ohlcvData[4], inTimePeriod)
+	ret := indicators.MFI(ohlcvData[2], ohlcvData[3], ohlcvData[4], ohlcvData[5], inTimePeriod)
 	for x := range ret {
 		r.Value = append(r.Value, &objects.Float{Value: math.Round(ret[x]*100) / 100})
 	}
